created encryption code for substitution cipher
program that is able to take a cipher alphabet to stdin and substitute those elements to there respective alphabetical characters in the string (ie message from user) and encrypt. Still have to implement decryption and merge with the teaser cipher. add commands in IO to allow user to choose ceaser or substitution method.